[test] Regroup and add new DP & device tests.
